+++ src/Storage/EverTask.Storage.Postgres/PostgresTaskStoreContext.cs
@@ -1,5 +1,4 @@
 namespace EverTask.Storage.Postgres;
 
-public class PostgresTaskStoreContext(
-    DbContextOptions<PostgresTaskStoreContext> options,
-    IOptions<ITaskStoreOptions> storeOptions) : TaskStoreEfDbContext<PostgresTaskStoreContext>(options, storeOptions);
+public class PostgresTaskStoreContext(DbContextOptions<PostgresTaskStoreContext> options)
+    : TaskStoreEfDbContext<PostgresTaskStoreContext>(options);