+++ src/Storage/EverTask.Storage.SqlServer/SqlServerTaskStoreContext.cs
@@ -1,5 +1,4 @@
-﻿namespace EverTask.Storage.SqlServer;
+namespace EverTask.Storage.SqlServer;
 
-public class SqlServerTaskStoreContext(
-    DbContextOptions<SqlServerTaskStoreContext> options,
-    IOptions<ITaskStoreOptions> storeOptions) : TaskStoreEfDbContext<SqlServerTaskStoreContext>(options, storeOptions);
+public class SqlServerTaskStoreContext(DbContextOptions<SqlServerTaskStoreContext> options)
+    : TaskStoreEfDbContext<SqlServerTaskStoreContext>(options);